changed double to long long, encountered many issues. project needs refactor

+++ src/demo/algorithm.h
@@ -7,5 +7,5 @@ public:
     virtual ~algorithm() = default;
 
     virtual void preprocessing(const std::vector<float>& points) = 0;
-    virtual std::vector<float> get_visibility(double x, double y) = 0;
+    virtual std::vector<float> get_visibility(float x, float y) = 0;
 };
